+++ src/components/promise/Header/style.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+// import { theme } from '@/styles/theme';
+
+export const Header = styled.header`
+  cursor: pointer;
+  flex-shrink: 0;
+`;
