+++ src/pages/promise/[id]/result/index.jsx
@@ -1,6 +1,6 @@
 import * as S from './style';
 
-const JoinSchedulePage = () => {
+const JoinResultPage = () => {
   return <S.Container>참여자 약속 장소 좋아요 페이지</S.Container>;
 };
-export default JoinSchedulePage;
+export default JoinResultPage;
