+++ src/pages/promise/[id]/schedule/index.jsx
@@ -1,6 +1,6 @@
 import * as S from './style';
 
-const ResultPage = () => {
+const JoinSchedulePage = () => {
   return <S.Container>참여자 스케줄 입력 페이지</S.Container>;
 };
-export default ResultPage;
+export default JoinSchedulePage;
