+++ src/pages/promise/create/date/index.jsx
@@ -1,6 +1,6 @@
 import * as S from './style';
 
-const SummaryPage = () => {
+const DatePage = () => {
   return <S.Container>약속 날짜 입력 페이지</S.Container>;
 };
-export default SummaryPage;
+export default DatePage;
