+++ src/pages/promise/create/info/index.jsx
@@ -1,6 +1,6 @@
 import * as S from './style';
 
-const SignInPage = () => {
+const InfoPage = () => {
   return <S.Container>약속 기본 정보 입력 페이지</S.Container>;
 };
-export default SignInPage;
+export default InfoPage;
