+++ src/pages/promise/create/location/index.jsx
@@ -1,6 +1,6 @@
 import * as S from './style';
 
-const InfoPage = () => {
+const LocationPage = () => {
   return <S.Container>약속 제안자 위치 입력 페이지</S.Container>;
 };
-export default InfoPage;
+export default LocationPage;
