+++ src/pages/promise/create/schedule/index.jsx
@@ -1,6 +1,6 @@
 import * as S from './style';
 
-const LocationPage = () => {
+const SchedulePage = () => {
   return <S.Container>약속 제안자 스케줄 입력 페이지</S.Container>;
 };
-export default LocationPage;
+export default SchedulePage;
