+++ src/pages/sign-in/index.jsx
@@ -1,6 +1,6 @@
 import * as S from './style';
 
-const SchedulePage = () => {
+const SignInPage = () => {
   return <S.Container>로그인 페이지</S.Container>;
 };
-export default SchedulePage;
+export default SignInPage;
