+++ src/pages/user/enter-schedule/index.jsx
@@ -1,6 +1,6 @@
 import * as S from './style';
 
-const UserPage = () => {
+const EnterSchedulePage = () => {
   return <S.Container>스케줄 입력 페이지</S.Container>;
 };
-export default UserPage;
+export default EnterSchedulePage;
