+++ src/pages/user/index/index.jsx
@@ -1,6 +1,6 @@
 import * as S from './style';
 
-const EnterSchedulePage = () => {
+const UserPage = () => {
   return <S.Container>유저 페이지</S.Container>;
 };
-export default EnterSchedulePage;
+export default UserPage;
